Change Mongodb write concern
The motivation for this change is to make writes to
Mongodb faster. This changes the write concern to the
default value "acknowledged" from majority. This should
enable Mongodb to handle writes faster, but might lead
to a greater risk of replication lag.

The env var is only used by the e2e tests:
https://github.com/alphagov/publishing-e2e-tests/blob/526b4b2edd88a6eeb6634d35c7186e51f2b265ca/docker-compose.yml