Demo of FindByPath tests without FactoryBot
These tests are pretty weird as they involve a real Mongoid model that
writes to the DB but is only relevant to this to this test file. I'm
sure there's got to be a better way to do these tests but the immediate
problem faced is a linting rule that about the use of a global constant
of CompatibleModel.

Using FactoryBot as a means to initialise CompatibleModel leaks the
knowledge of this model out into the wider application - which seems
sub-optimal for unit tests. The approach I've taken to resolve this is
to handle the FactoryBot logic within the class itself so that it is no
longer needed.